+++ src/config.py
@@ -1,0 +1,7 @@
+from pathlib import Path
+
+REPO_NAME = "ClimatePolicyRadar/all-document-text-data"
+REPO_URL = f"https://huggingface.co/datasets/{REPO_NAME}"
+CACHE_DIR = Path(__file__).parent / "../cache"
+
+REVISION = "main"  # Use this to set a commit hash. Recommended!
