+++ src/data/earth_vectors/ne_50m_admin_0_countries.prj
